Split controllers - new LogController
