+++ Controller/DashboardController.php
@@ -34,7 +34,7 @@ class DashboardController extends Controller
         $form = $this->createForm(new CronType(), new Cron());
 
         return $this->render('FOACronBundle:Dashboard:index.html.twig', [
-            'crons' => $cronManager->get(),
+            'crons' => $cronManager->getCrons(),
             'raw'   => $cronManager->getRaw(),
             'form'  => $form->createView(),
         ]);
@@ -65,7 +65,7 @@ class DashboardController extends Controller
         }
 
         return $this->render('FOACronBundle:Dashboard:index.html.twig', [
-            'crons' => $cronManager->get(),
+            'crons' => $cronManager->getCrons(),
             'raw'   => $cronManager->getRaw(),
             'form'  => $form->createView(),
         ]);
@@ -81,7 +81,7 @@ class DashboardController extends Controller
     public function editAction($id)
     {
         $cronManager = new CronManager();
-        $cronList = $cronManager->get();
+        $cronList = $cronManager->getCrons();
         $this->addFlash('message', $cronManager->getOutput());
         $this->addFlash('error', $cronManager->getError());
         $form = $this->createForm(new CronType(), $cronList[$id]);
@@ -139,7 +139,7 @@ class DashboardController extends Controller
     protected function suspendTask($id, $state)
     {
         $cronManager = new CronManager();
-        $cronList = $cronManager->get();
+        $cronList = $cronManager->getCrons();
         $this->addFlash('message', $cronManager->getOutput());
         $this->addFlash('error', $cronManager->getError());
 
@@ -168,33 +168,6 @@ class DashboardController extends Controller
         $this->addFlash('error', $cronManager->getError());
 
         return $this->redirect($this->generateUrl('foa_cron_index'));
-    }
-
-    /**
-     * Gets a log file
-     *
-     * @param $id   - the line of the cron in the cron table
-     * @param $type - the type of file, log or error
-     *
-     * @return Response
-     */
-    public function fileAction($id, $type)
-    {
-        $cronManager = new CronManager();
-        $cronList = $cronManager->get();
-
-        /**
-         * @var Cron $cron
-         */
-        $cron = $cronList[$id];
-
-        $data = [];
-        $data['file'] = ($type == 'log') ? $cron->getLogFile() : $cron->getErrorFile();
-        $data['content'] = file_get_contents($data['file']);
-
-        $serializer = new Serializer([], ['json' => new JsonEncoder()]);
-
-        return new Response($serializer->serialize($data, 'json'));
     }
 
     /**
